Update JLDiff.py
In order to pass HTML validation, you need a doctype declaration. Extraneous/inferred elements are unnecessary and not mandated by HTML5. This HTML5 syntax validates W3C and works in all browsers. Win-win!
The only obsolete tag is `font`, but it is easy to replace by `span` with a `class` which will make future style changes/enhancements easier too.